Redoing the code for api, previous version didn't work and theree was too much to debug, working on peices at a time now that I can test it.
